Fixed text size method. Added create text block.
Fixed the text size method to return the correct size. Added a method to
take a string of text and turn it into a text block of a certain width.
Currently a little buggy with new line characters, and it may go over
the width by as much as 5 (highest seen error).